+++ src/get_text/iterm.rs
@@ -1,0 +1,27 @@
+use std::{env, process::Command};
+
+use crate::get_text::find_command_output;
+
+/// This is an AppleScript that gets the contents of the current session
+const SCRIPT: &str = r#"
+tell application "iTerm2"
+    tell current session of current window
+        contents
+    end tell
+end tell
+"#;
+
+pub fn get_text_iterm(cmd: &str, depth: usize) -> Option<String> {
+    if env::var("ITERM_SESSION_ID").is_err() {
+        return None;
+    }
+
+    log::debug!("getting the command output from iTerm");
+
+    let output = Command::new("osascript")
+        .args(["-e", SCRIPT])
+        .output()
+        .ok()?;
+
+    find_command_output(cmd, output.stdout, Some(depth))
+}